requirements.txt
# -*- coding: utf-8 -*-
torch==1.1.0
tensorflow==1.14.0
stanfordnlp==0.2.0
numpy==1.17.4
cython==0.29.14
# benepar==0.1.2
spacy==2.2.2
nltk==3.4.5
PyStanfordDependencies==0.3.1
pyyaml==5.1.2
pyconll==2.2.1
conllu==2.2

